feat: seo 개선 (#89)
* feat: 서치 콘솔 인증 파일 추가

* feat: 오픈그래프 메타데이터 추가

* feat: 구글 애널리틱스 추가

* fix: 함수 컨벤션에 맞게 수정 및 클라리티 스크립트 파일명 수정
