+++ client/src/apis/repository/room.repository.ts
@@ -1,5 +1,5 @@
-import { addAuthHeader, baseClient, createAuthHeader } from "@/modules/fetchClient";
-import { ICreateRoomResponse, IPatchRoom, IRoomModel } from "@/types/room.type";
+import { baseClient, createAuthHeader } from "@/modules/fetchClient";
+import { IPatchRoom, IRoomModel } from "@/types/room.type";
 
 export const getRoomInfo = async (roomId: string, accessToken?: string) => {
     const headers = createAuthHeader(accessToken);
