+++ client/src/app/googleAnalytics.tsx
@@ -1,0 +1,18 @@
+import Script from "next/script";
+
+export const GoogleAnalytics = () => {
+    return (
+        <>
+            <Script async src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_ID}`} strategy="afterInteractive" />
+            <Script id="google-analytics" strategy="afterInteractive">
+                {`window.dataLayer = window.dataLayer || [];
+                function gtag(){dataLayer.push(arguments);}
+                gtag('js', new Date());
+
+                gtag('config', '${process.env.NEXT_PUBLIC_GA_ID}');`}
+            </Script>
+        </>
+    );
+};
+
+export default GoogleAnalytics;
