+++ client/src/app/lobby/page.tsx
@@ -4,6 +4,11 @@ import { QUERY_KEY } from "@/constants/queryKey.const";
 import { HydrationBoundary, QueryClient, dehydrate } from "@tanstack/react-query";
 import { getToken } from "@/utils/getToken.utill";
 import { formatRoomsData } from "../../apis/service/room.service";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+    title: "로비",
+};
 
 const Lobby = async () => {
     const queryClient = new QueryClient();
